Fix an issue with events imported from Github Archive
Events from Github Archive are exported via Google BigQuery. These
do not create valid JSON. The format is a list of JSON objects but
not comma separated. The previous logic to join these stripped off
outer curly braces and then joined them, leading to a garbled CSV.

In addition, the extra fields from the payload were also not being
processed correctly. Only the first field in the list was getting
handled. Added the "items()" call to the decoded values to fix it.

Closes Issue #62

Signed-off-by: Augustina Ragwitz <augustina.ragwitz@ibm.com>